requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-querycount==0.8.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyparsing==3.2.3
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
pre-commit